Breaks steps into individual components. Uses nested routing
