+++ app/routes/user/streaming_steps/index.tsx
@@ -1,0 +1,7 @@
+export default function StreamingSteps() {
+  return (
+    <div className="usa-prose">
+      <p>Keys are bound to your account.</p>
+    </div>
+  )
+}
